# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Реактив CRM",
    "description": "CRM продаж и закупок химической продукции. Настраиваемые финансовые профили, права и шаблоны. Количества и деньги — десятичные строки.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/counterparties": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Clients",
        "operationId": "clients_api_v1_counterparties_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name",
              "title": "Sort"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc",
              "title": "Direction"
            }
          },
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Clients Api V1 Counterparties Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      },
      "post": {
        "tags": [
          "CRM"
        ],
        "summary": "Create Client",
        "operationId": "create_client_api_v1_counterparties_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Client Api V1 Counterparties Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/counterparties/{entity_id}": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Client Detail",
        "operationId": "client_detail_api_v1_counterparties__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Client Detail Api V1 Counterparties  Entity Id  Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      },
      "patch": {
        "tags": [
          "CRM"
        ],
        "summary": "Edit Client",
        "operationId": "edit_client_api_v1_counterparties__entity_id__patch",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Edit Client Api V1 Counterparties  Entity Id  Patch"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/counterparties/{entity_id}/contacts": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Contacts",
        "operationId": "contacts_api_v1_counterparties__entity_id__contacts_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Contacts Api V1 Counterparties  Entity Id  Contacts Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      },
      "post": {
        "tags": [
          "CRM"
        ],
        "summary": "Create Contact",
        "operationId": "create_contact_api_v1_counterparties__entity_id__contacts_post",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Contact Api V1 Counterparties  Entity Id  Contacts Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
...
    "/api/v1/tasks/{entity_id}": {
      "patch": {
        "tags": [
          "CRM"
        ],
        "summary": "Edit Task",
        "operationId": "edit_task_api_v1_tasks__entity_id__patch",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Edit Task Api V1 Tasks  Entity Id  Patch"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClientInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 250,
            "minLength": 1,
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "client",
              "supplier",
              "both"
            ],
            "title": "Kind",
            "default": "client"
          },
          "country": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "tax_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 250
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ClientInput"
      },
...
    }
  }
}